chore: move tests to AWS
