+++ prowler/providers/azure/lib/mutelist/mutelist.py
@@ -9,12 +9,10 @@ class MutelistAzure(Mutelist):
         self,
         finding: Any,
         subscription: str,
-        check_id: str,
     ) -> bool:
         return self.is_muted(
-            self._mutelist,
             subscription,
-            check_id,
+            finding.check_metadata.CheckID,
             # TODO: add region to the findings when we add Azure Locations
             # finding.region,
             "",
