+++ tests/providers/aws/lib/mutelist/fixtures/aws_mutelist.yaml
